+++ calMaCY/calMaCY.pyx
@@ -13,7 +13,7 @@ cdef np.ndarray[double, ndim=1] _calMaCY(np.ndarray[long, ndim=1] d, long window
     adjust = int(window/2)
     m = np.zeros(len(d),dtype=np.float64)
     for i in range(len(d)-window):
-        m[i+adjust] = np.sum(d[i:i+window])/window
+        m[i+adjust] = np.sum(d[i:i+window], dtype=np.int64)/window
     return m
 
 def calMaCY(d, window):
